backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.15.0
